Product Collection: move "Display Settings" from Toolbar to Inspector Controls (#52666)
* Move display layout menu from toolbar to inspector controls

This commit moves the display layout menu from the toolbar to the inspector
controls panel. The display settings include:

- Products per page
- Offset
- Max pages to show

The display settings are only shown when the query is not inherited from the
template.

* Delete display settings toolbar

The display settings toolbar has been removed as these controls are being moved
to the inspector panel for better UX.

* Add changefile(s) from automation for the following project(s): woocommerce-blocks

* E2E Tests: Move display layout menu from toolbar to inspector controls

The display layout menu has been moved from the toolbar to the inspector controls.
This commit updates the tests to reflect this change.

* Fix failing test

* Move display settings to Settings section in inspector controls

- Move display controls to "Settings" section
- Keep "Offset" and "Max pages to show" hidden
- Place "Products per page" under "Columns" control as they're related.

* Extract query reset logic from getDefaultSettings into a separate function

* Remove Dimensions section and move width control to Settings section

* Make Products per Page and Columns input boxes consistent in size

The Products per Page and Columns input boxes had different sizes in the inspector controls. This commit makes them consistent by adding the __next40pxDefaultSize prop to the Columns RangeControl component to match the Products per Page control.

* Move range controls next to each other in inspector controls

Reorders the inspector controls to group similar control types together,
specifically placing the range controls (Products per page and Columns)
adjacent to each other for better UX and information organization.

* Fix E2E test

---------

Co-authored-by: github-actions <github-actions@github.com>